Split SerialiseRaw module out of Cardano.Api.Typed
